requirements.txt
django
djangorestframework
django-cors-headers
Setuptools<81
coreapi
drf-yasg
django-rest-passwordreset